requirements.txt
python-dotenv==1.0.0
rich==14.2.0
requests>=2.31.0
docker